# Improve template mechanisms