* Delete `Eq`, `Gt`, `appendIfNotPresent` and `trim` functions
* Add `[flags]` in UseLine
* Simplify other functions
* Simplify templates

⚠️   **Public API changes:** Delete public `Eq`, `Gt` functions.
*`Eq` is used nowhere, `Gt` was used only in template.* I looked in some projects that use cobra, and I found out, what (almost) no one uses them.

---
Minor performance improvement.
Benchmark for command with 4 flags and one child command:
```
benchmark                   old ns/op     new ns/op     delta
BenchmarkCmdUsageFunc-4     335860        319290        -4.93%

benchmark                   old allocs     new allocs     delta
BenchmarkCmdUsageFunc-4     562            543            -3.38%

benchmark                   old bytes     new bytes     delta
BenchmarkCmdUsageFunc-4     21623         21037         -2.71%
```

/cc @spf13 @eparis