@fabianofranz 2017-05-09
@bogem @eparis We should take better care in removing publicly exported functions (`Eq` and `Gt` in this case). `kubectl` in [Kubernetes](https://github.com/kubernetes/kubernetes/) uses it.