@bogem 2017-05-09
@fabianofranz can you please show me the code, where these functions are used