@fabianofranz 2017-05-09
@bogem we write our own templates, examples:

https://github.com/kubernetes/kubernetes/blob/master/pkg/kubectl/cmd/templates/templater.go#L102-L103
https://github.com/kubernetes/kubernetes/blob/master/pkg/kubectl/cmd/templates/templates.go#L34