@bogem 2017-05-09
You can easily delete `eq` from `templateFuncs`, because it's used [nowhere](https://github.com/kubernetes/kubernetes/search?l=Go&q=eq&type=&utf8=%E2%9C%93) in kubernetes at all.