@bogem 2017-05-09
`gt` is used only [here](https://github.com/kubernetes/kubernetes/blob/c19569f03f09eb4e8638d8f6a20b0a7a751abcc2/pkg/kubectl/cmd/templates/templates.go#L34), but you can rewrite it with standard function from text/template [in this way](https://github.com/spf13/cobra/blob/master/command.go#L368): `{{if gt (len .Aliases) 0}}Aliases:`.