@fabianofranz 2017-05-09
Sure, fixing would have been simple. But we rely on [semver](http://semver.org/) and we also have users (e.g. OpenShift) which may be using them and we can't break anyone. API changes should only happen in major release.