@bogem 2017-05-09
`(templater *templater) templateFuncs` is a private function, so changing that will not harm to API AFAIK. 
And the behaviour of `SectionAliases` will be fully same, if you will replace `gt .Aliases 0` with `gt (len .Aliases) 0`