@fabianofranz 2017-05-09
Thanks but no. Kube cmd users could be calling `SetHelpTemplate` and `SetUsageTemplate`.